moved my PlaneType to ..model/enums
