+++ airport-api/src/main/java/cz/muni/airport/dto/AirplaneDTO.java
@@ -1,7 +1,7 @@
 package cz.muni.airport.dto;
 
 import java.util.List;
-import cz.muni.airport.model.PlaneType;
+import cz.muni.airport.model.enums.PlaneType;
 import java.util.Objects;
 
 /**
