+++ airport-dao/src/main/java/cz/muni/airport/model/enums/PlaneType.java
@@ -1,4 +1,4 @@
-package cz.muni.airport.model;
+package cz.muni.airport.model.enums;
 
 /**
  *
